+++ erpnext/patches/v12_0/move_target_distribution_from_parent_to_child.py
@@ -16,3 +16,7 @@ def execute():
                 `tab{child_doc}`.parent = `tab{parent_doc}`.name
                 and `tab{parent_doc}`.distribution_id is not null and `tab{parent_doc}`.distribution_id != ''
         """.format(parent_doc = d, child_doc = "Target Detail"))
+
+    frappe.delete_doc("Report", "Sales Partner-wise Transaction Summary")
+    frappe.delete_doc("Report", "Sales Person Target Variance Item Group-Wise")
+    frappe.delete_doc("Report", "Territory Target Variance Item Group-Wise")
