+++ erpnext/selling/report/sales_partner_target_variance_based_on_item_group/sales_partner_target_variance_based_on_item_group.js
@@ -1,7 +1,8 @@
-// Copyright (c) 2015, Frappe Technologies Pvt. Ltd. and Contributors
-// License: GNU General Public License v3. See license.txt
+// Copyright (c) 2016, Frappe Technologies Pvt. Ltd. and contributors
+// For license information, please see license.txt
+/* eslint-disable */
 
-frappe.query_reports["Territory Target Variance Item Group-Wise"] = {
+frappe.query_reports["Sales Partner Target Variance based on Item Group"] = {
 	"filters": [
 		{
 			fieldname:"company",
