+++ erpnext/selling/report/sales_partner_target_variance_based_on_item_group/sales_partner_target_variance_based_on_item_group.json
@@ -1,19 +1,21 @@
 {
  "add_total_row": 0, 
- "apply_user_permissions": 1, 
- "creation": "2013-06-21 12:15:00", 
+ "creation": "2019-03-25 18:22:37.323995", 
+ "disable_prepared_report": 0, 
  "disabled": 0, 
  "docstatus": 0, 
  "doctype": "Report", 
- "idx": 3, 
+ "idx": 0, 
  "is_standard": "Yes", 
- "modified": "2017-02-24 20:13:29.705321", 
+ "letter_head": "Gadgets International", 
+ "modified": "2019-03-25 18:22:37.323995", 
  "modified_by": "Administrator", 
  "module": "Selling", 
- "name": "Territory Target Variance Item Group-Wise", 
+ "name": "Sales Partner Target Variance based on Item Group", 
  "owner": "Administrator", 
+ "prepared_report": 0, 
  "ref_doctype": "Sales Order", 
- "report_name": "Territory Target Variance Item Group-Wise", 
+ "report_name": "Sales Partner Target Variance based on Item Group", 
  "report_type": "Script Report", 
  "roles": [
   {
