+++ erpnext/selling/report/sales_partner_transaction_summary/sales_partner_transaction_summary.json
@@ -1,6 +1,6 @@
 {
- "add_total_row": 0, 
- "creation": "2019-03-15 17:42:00.631020", 
+ "add_total_row": 1, 
+ "creation": "2019-03-25 18:15:09.920739", 
  "disable_prepared_report": 0, 
  "disabled": 0, 
  "docstatus": 0, 
@@ -8,14 +8,14 @@
  "idx": 0, 
  "is_standard": "Yes", 
  "letter_head": "Gadgets International", 
- "modified": "2019-03-15 17:42:00.631020", 
+ "modified": "2019-03-25 18:15:09.920739", 
  "modified_by": "Administrator", 
  "module": "Selling", 
- "name": "Sales Partner Target Variance Item Group-Wise", 
+ "name": "Sales Partner Transaction Summary", 
  "owner": "Administrator", 
  "prepared_report": 0, 
  "ref_doctype": "Sales Order", 
- "report_name": "Sales Partner Target Variance Item Group-Wise", 
+ "report_name": "Sales Partner Transaction Summary", 
  "report_type": "Script Report", 
  "roles": [
   {
