+++ erpnext/selling/report/sales_partner_transaction_summary/sales_partner_transaction_summary.py
