+++ erpnext/selling/report/sales_person_target_variance_based_on_item_group/sales_person_target_variance_based_on_item_group.js
@@ -1,7 +1,8 @@
-// Copyright (c) 2015, Frappe Technologies Pvt. Ltd. and Contributors
-// License: GNU General Public License v3. See license.txt
+// Copyright (c) 2016, Frappe Technologies Pvt. Ltd. and contributors
+// For license information, please see license.txt
+/* eslint-disable */
 
-frappe.query_reports["Sales Person Target Variance Item Group-Wise"] = {
+frappe.query_reports["Sales Person Target Variance Based On Item Group"] = {
 	"filters": [
 		{
 			fieldname:"company",
