+++ erpnext/selling/report/sales_person_target_variance_based_on_item_group/sales_person_target_variance_based_on_item_group.json
@@ -1,19 +1,21 @@
 {
  "add_total_row": 0, 
- "apply_user_permissions": 1, 
- "creation": "2013-06-21 12:14:15", 
+ "creation": "2019-03-25 22:16:49.040998", 
+ "disable_prepared_report": 0, 
  "disabled": 0, 
  "docstatus": 0, 
  "doctype": "Report", 
- "idx": 3, 
+ "idx": 0, 
  "is_standard": "Yes", 
- "modified": "2017-02-24 20:14:50.589149", 
+ "letter_head": "Gadgets International", 
+ "modified": "2019-03-25 22:16:49.040998", 
  "modified_by": "Administrator", 
  "module": "Selling", 
- "name": "Sales Person Target Variance Item Group-Wise", 
+ "name": "Sales Person Target Variance Based On Item Group", 
  "owner": "Administrator", 
+ "prepared_report": 0, 
  "ref_doctype": "Sales Order", 
- "report_name": "Sales Person Target Variance Item Group-Wise", 
+ "report_name": "Sales Person Target Variance Based On Item Group", 
  "report_type": "Script Report", 
  "roles": [
   {
