+++ erpnext/selling/report/sales_person_target_variance_based_on_item_group/sales_person_target_variance_based_on_item_group.py
@@ -3,10 +3,9 @@
 
 from __future__ import unicode_literals
 import frappe
-from erpnext.selling.report.sales_partner_target_variance_item_group_wise.item_group_wise_sales_target_variance import get_data_column
+from erpnext.selling.report.sales_partner_target_variance_based_on_item_group.item_group_wise_sales_target_variance import get_data_column
 
 def execute(filters=None):
 	data = []
 
-	return get_data_column(filters, "Territory")
-
+	return get_data_column(filters, "Sales Person")
