+++ erpnext/selling/report/territory_target_variance_based_on_item_group/territory_target_variance_based_on_item_group.js
@@ -2,7 +2,7 @@
 // For license information, please see license.txt
 /* eslint-disable */
 
-frappe.query_reports["Sales Partner Target Variance Item Group-Wise"] = {
+frappe.query_reports["Territory Target Variance Based On Item Group"] = {
 	"filters": [
 		{
 			fieldname:"company",
