+++ erpnext/selling/report/territory_target_variance_based_on_item_group/territory_target_variance_based_on_item_group.json
@@ -1,6 +1,6 @@
 {
- "add_total_row": 1, 
- "creation": "2019-03-15 16:21:16.088831", 
+ "add_total_row": 0, 
+ "creation": "2019-03-25 22:20:59.033199", 
  "disable_prepared_report": 0, 
  "disabled": 0, 
  "docstatus": 0, 
@@ -8,14 +8,14 @@
  "idx": 0, 
  "is_standard": "Yes", 
  "letter_head": "Gadgets International", 
- "modified": "2019-03-15 16:21:16.088831", 
+ "modified": "2019-03-25 22:20:59.033199", 
  "modified_by": "Administrator", 
  "module": "Selling", 
- "name": "Sales Partner-wise Transaction Summary", 
+ "name": "Territory Target Variance Based On Item Group", 
  "owner": "Administrator", 
  "prepared_report": 0, 
  "ref_doctype": "Sales Order", 
- "report_name": "Sales Partner-wise Transaction Summary", 
+ "report_name": "Territory Target Variance Based On Item Group", 
  "report_type": "Script Report", 
  "roles": [
   {
